# [CBR 7.9] perf: Disallow mis-matched inherited group reads